requirements.txt
alabaster==0.7.12
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
awscli==1.19.7
Babel==2.9.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
botocore==1.20.7
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coverage==5.4
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
flake8==3.8.4
gitdb==4.0.5
GitPython==3.1.13
idna==2.10
imagesize==1.2.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.6.1
pyflakes==2.2.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
streamlit==0.76.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
